refactor: move article header into separate component
