+++ apps/marketing/src/components/blog/BlogArticle.tsx
